pyproject.toml
[project]
name = "dream-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "taketo44",email = "take1044da@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
